feat(ui): fix HTML parse_mode serialization, refine card layout, and add high-res progress bar (#41)
* feat(ui): refine Telegram card layout and add high-resolution progress bar

* fix(telegram): enable encodeDefaults so parse_mode HTML is included in API payloads

* feat(ui): display download progress percentage with two decimal places

+++ src/main/kotlin/app/hononeko/notifier/adapter/outbound/qbittorrent/QBittorrentClientAdapter.kt
@@ -150,7 +150,7 @@ class QBittorrentClientAdapter(
             torrentList.firstOrNull { it.hash.equals(hash, ignoreCase = true) }
                 ?: return Either.Right(null)
 
-        val progressPercent = (torrentDto.progress * 100).toInt().coerceIn(0, 100)
+        val progressPercent = (torrentDto.progress * 100.0).coerceIn(0.0, 100.0)
         val state = mapState(torrentDto.state)
 
         return Either.Right(
